Read RFC 2231 continuations in one pass and honor every charset the platform names
`extendedContinuation` looked each numbered section up with
`extractHeaderParam`, which tokenizes the whole header on every call, so a
sender who writes N legal `filename*N*=` sections made the parser do work
that grows with N². Measured with `swiftc -O`: 1,024 sections took 0.5 s,
4,096 took 8 s, and 8,192 did not finish inside 25 s, where the previous
release read the same headers in tens of milliseconds. The header is now
tokenized once into `parameters(of:)`, and both the single-attribute lookup
and the continuation collector read from that list. The collector also
applies RFC 2231 §3 as written: sections count in decimal from 0 without
leading zeroes or gaps, and the first occurrence of a repeated section
stands, matching `extractHeaderParam`.

`decodeExtendedBytes` whitelisted three charsets and returned nil for every
other label, including the blank one. RFC 2231 §4 makes the charset field
optional with the `'` delimiters present, so `filename*=''invoice.pdf` is a
legal way to write `invoice.pdf`, and a filename labelled `windows-1252` or
`iso-8859-15` is ordinary real-world mail; all of them lost the name and
fell through to a generated one. A blank charset now decodes as UTF-8, which
reads every US-ASCII value unchanged, and any other label resolves through
`String.Encoding(ianaCharsetName:)`, the same lookup the encoded-word
decoder already uses. A label the platform cannot name still yields nil,
and the literal spelling of the parameter is used instead.

The RFC 2231 helpers move to `EMLParser+RFC2231.swift`; there is no
behaviour change in the move beyond the two above.

Verification: `swift test` passes 624 tests in 77 suites; `swiftlint lint
--strict` reports no violations. Five tests are added: blank charset,
platform-named charsets (windows-1252, ISO-8859-1, and ISO-8859-15 on Darwin only:
swift-corelibs-foundation names that charset but has no converter for it), unknown
charset falling back to the literal spelling, section numbering rules, and a
4,096-section header that must decode in under two seconds. The blank-charset,
platform-charset and linear-work tests fail on the parent commit (the last one
after 34 s in a debug build); the other two pin behaviour that already held.
