+++ Tests/SwiftIMAPTests/EMLTests.swift
@@ -326,6 +326,64 @@ struct MIMEParameterParsingTests {
         #expect(EMLParser.cleanContentType(contentType) == "application/pdf")
     }
 
+    @Test("An extended parameter may leave the charset field blank")
+    func extendedParameterWithBlankCharsetIsDecoded() {
+        // RFC 2231 §4: "it is perfectly permissible to leave either the
+        // character set or language field blank", the `'` delimiters staying.
+        let header = "attachment; filename*=''invoice.pdf"
+
+        #expect(EMLParser.extractFilename(from: header) == "invoice.pdf")
+    }
+
+    @Test("An extended parameter decodes in any charset the platform names")
+    func extendedParameterHonorsPlatformCharsets() {
+        #expect(EMLParser.extractFilename(from: "attachment; filename*=windows-1252''invoice.pdf") == "invoice.pdf")
+        #expect(EMLParser.extractFilename(from: "attachment; filename*=ISO-8859-1''caf%E9.txt") == "café.txt")
+        #if canImport(Darwin)
+        // swift-corelibs-foundation names ISO-8859-15 but has no converter for
+        // it, so on Linux this value decodes to nil and the literal spelling is
+        // used instead; only the Darwin converter set reaches the euro sign.
+        #expect(EMLParser.extractFilename(from: "attachment; filename*=iso-8859-15''%A4.txt") == "€.txt")
+        #endif
+    }
+
+    @Test("An extended parameter in an unknown charset yields the literal spelling")
+    func extendedParameterUnknownCharsetFallsBackToLiteral() {
+        let header = "attachment; filename=\"fallback.pdf\"; filename*=x-no-such-charset''a.pdf"
+
+        #expect(EMLParser.extractFilename(from: header) == "fallback.pdf")
+    }
+
+    @Test("Continuation sections end at the first gap and reject leading zeroes")
+    func continuationSectionsAreContiguousDecimals() {
+        // RFC 2231 §3: "neither leading zeroes nor gaps in the sequence are
+        // allowed" — `*01*` is not section 1, and section 2 is unreachable
+        // without it.
+        #expect(EMLParser.extractFilename(from: "application/pdf; name*0*=UTF-8''a; name*01*=b; name*2*=c") == "a")
+        #expect(EMLParser.extractFilename(from: "application/pdf; name*0*=UTF-8''a; name*2*=c") == "a")
+        // The first occurrence of a repeated section stands.
+        #expect(EMLParser.extractFilename(from: "application/pdf; name*0*=UTF-8''a; name*1*=b; name*1*=z") == "ab")
+    }
+
+    @Test("Reading a parameter with many continuation sections stays linear")
+    func manyContinuationSectionsStayLinear() {
+        // A sender chooses the section count. Each section used to be found
+        // by re-tokenizing the whole header, so 4096 legal sections cost
+        // seconds; one pass over the parameters costs milliseconds.
+        let sections = 4096
+        let header = "attachment; filename*0*=UTF-8''a; "
+            + (1..<sections).map { "filename*\($0)*=a" }.joined(separator: "; ")
+
+        let clock = ContinuousClock()
+        var filename: String?
+        let elapsed = clock.measure {
+            filename = EMLParser.extractFilename(from: header)
+        }
+
+        #expect(filename == String(repeating: "a", count: sections))
+        #expect(elapsed < .seconds(2))
+    }
+
     @Test("An extended name parameter reads back from a Content-Type")
     func extendedNameParameterReadsBack() {
         let contentType = "application/pdf; name*=UTF-8''%EB%B0%9C%ED%91%9C.pdf"
